refactor: PR feedback
