+++ src/pages/common/Breadcrumbs/Breadcrumbs.tsx
@@ -60,14 +60,6 @@ const generateSteps = (content: Content, variant: Variant) => {
     case 'news':
       const news = content as News
       steps.push({ text: 'News', link: '/news' })
-
-      // if (news.category) {
-      //   steps.push({
-      //     text: news.category.name,
-      //     link: `/news?category=${news.category.id}`,
-      //   })
-      // }
-
       steps.push({ text: news.title })
       break
   }
