+++ src/routes/_.news.$slug.edit.tsx
@@ -1,10 +1,9 @@
 import { useLoaderData } from '@remix-run/react'
 import { News } from 'oa-shared'
-import { AuthRoute } from 'src/pages/common/AuthRoute'
 import { NewsEdit } from 'src/pages/News/NewsEdit'
 import { createSupabaseServerClient } from 'src/repository/supabase.server'
 import { newsServiceServer } from 'src/services/newsService.server'
-import { tagsService } from 'src/services/tagsService'
+import { tagsServiceServer } from 'src/services/tagsService.server'
 
 import type { LoaderFunctionArgs } from '@remix-run/node'
 import type { DBNews } from 'oa-shared'
@@ -23,7 +22,7 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
   }
 
   const dbNews = result.data as unknown as DBNews
-  const tags = await tagsService.getTags(client, dbNews.tags)
+  const tags = await tagsServiceServer.getTags(client, dbNews.tags)
   const heroImage = await newsServiceServer.getHeroImage(
     client,
     dbNews.hero_image,
@@ -37,9 +36,5 @@ export default function Index() {
   const data = useLoaderData<typeof loader>()
   const news = data.news as News
 
-  return (
-    <AuthRoute>
-      <NewsEdit news={news} />
-    </AuthRoute>
-  )
+  return <NewsEdit news={news} />
 }
