+++ src/routes/_.questions.$slug.edit.tsx
@@ -1,7 +1,6 @@
 import { useLoaderData } from '@remix-run/react'
 import { Question } from 'oa-shared'
 import { IMAGE_SIZES } from 'src/config/imageTransforms'
-import { AuthRoute } from 'src/pages/common/AuthRoute'
 import { QuestionEdit } from 'src/pages/Question/QuestionEdit'
 import { createSupabaseServerClient } from 'src/repository/supabase.server'
 import { questionServiceServer } from 'src/services/questionService.server'
@@ -42,9 +41,5 @@ export default function Index() {
   const data = useLoaderData<typeof loader>()
   const question = data.question as Question
 
-  return (
-    <AuthRoute>
-      <QuestionEdit question={question} />
-    </AuthRoute>
-  )
+  return <QuestionEdit question={question} />
 }
