+++ src/routes/api.questions.ts
@@ -4,9 +4,9 @@ import { IModerationStatus, Question } from 'oa-shared'
 import { ITEMS_PER_PAGE } from 'src/pages/Question/constants'
 import { createSupabaseServerClient } from 'src/repository/supabase.server'
 import { discordServiceServer } from 'src/services/discordService.server'
+import { utilsServiceServer } from 'src/services/utilsService.server'
+import { validateImages } from 'src/utils/helpers'
 import { convertToSlug } from 'src/utils/slug'
-
-import { isDuplicateNewSlug, validateImages } from './utils'
 
 import type { LoaderFunctionArgs } from '@remix-run/node'
 import type { SupabaseClient, User } from '@supabase/supabase-js'
@@ -119,7 +119,9 @@ export const action = async ({ request }: LoaderFunctionArgs) => {
 
     const slug = convertToSlug(data.title)
 
-    if (await isDuplicateNewSlug(slug, client, 'questions')) {
+    if (
+      await utilsServiceServer.isDuplicateNewSlug(slug, client, 'questions')
+    ) {
       return Response.json(
         {},
         {
