+++ src/services/tagsService.ts
@@ -1,6 +1,4 @@
-import { Tag } from 'oa-shared'
-
-import type { SupabaseClient } from '@supabase/supabase-js'
+import type { Tag } from 'oa-shared'
 
 const getAllTags = async () => {
   try {
@@ -12,24 +10,6 @@ const getAllTags = async () => {
   }
 }
 
-const getTags = async (client: SupabaseClient, tagIds: number[]) => {
-  let tags: Tag[] = []
-
-  if (tagIds?.length > 0) {
-    const tagsResult = await client
-      .from('tags')
-      .select('id,name,created_at,modified_at')
-      .in('id', tagIds)
-
-    if (tagsResult.data) {
-      tags = tagsResult.data.map((x) => Tag.fromDB(x))
-    }
-  }
-
-  return tags
-}
-
 export const tagsService = {
   getAllTags,
-  getTags,
 }
